@@ -1,18 +1,17 @@
 import 'package:flutter/material.dart';
 
-class ConnectaxMediumHeader extends StatelessWidget {
-  const ConnectaxMediumHeader({Key? key, required this.text}) : super(key: key);
-  final String text;
+class ConnectaxMediumHeader extends Text {
+  const ConnectaxMediumHeader({
+    required String text,
+    super.key,
+    super.textAlign = TextAlign.left,
+    super.style = const TextStyle(fontSize: 25, fontWeight: FontWeight.w500),
+  }) : super(text);
 
   @override
-  Widget build(BuildContext context) {
-    return Padding(
+  Widget build(BuildContext context) => Padding(
       padding: const EdgeInsets.symmetric(vertical: 30),
-      child: Text(
-        text,
-        textAlign: TextAlign.left,
-        style: const TextStyle(fontSize: 25, fontWeight: FontWeight.w500),
-      ),
+        child: super.build(context),
     );
-  }
+  
 }
